Class: Vuex for everyone Chapter: 2 Lesson: 6 Title: Vuex actions
In this lesson we implemented a fetchProducts action and added a loading spinner to the products page to display when our loading variable is true. I tried to implement this with async/await but it does not appear to work despite documentation indicating it should. I will probably revisit this.