Class: Vuex for everyone Chapter: 3 Lesson: 8 Title: Add products to the cart Link: https://vueschool.io/lessons/add-products-to-the-cart
In this lesson we put together an action for adding products to the cart and set up mutations to keep the products inventory up to date as well as the item quantities in the cart up to date. We added a button for each item to add it to the cart. It functions as expected but on refresh everything is reset(I suspect this is the intention).