Class: Vuex for everyone Chapter: 3 Lesson: 9 Title: Vuex mutation history and vue devtools Link: https://vueschool.io/lessons/vuex-mutation-history-and-vue-devtools
In this course we went over the value added by having vue devtools installed when it comes to debugging live code with mutations and being able to view everything in the vuex state as well as being able to travel back and forth in time(the callstack) to see what the state was in those points in time as well as being able to see what mutation calls were made, what their parameters were and what their effect was. Very useful!