Class: Vuex for everyone Chapter: 3 Lesson: 10 Title: Cart items and total Link: https://vueschool.io/lessons/cart-items-and-total
Summary: In this lesson we added a shopping cart component to give a view into the shopping cart to the user, adding price, quantity and name as well as total eventually. We used a Vue.filter to add currency to the displayed values. I ran into a hiccup on my end for this part due to a typo that was not clear to debug when creating the shopping cart component({{proguct.title}} instead of {{product.title}}). This took me a bit to debug and find the actual cause of my shopping cart not displaying.