Class: Vuex for everyone Chapter: 4 Lesson: 12 Title: Dynamig vuex getters Link: https://vueschool.io/lessons/dynamic-vuex-getters
In this lesson we learned that though getters do not typically accept arguments, you can use currying to make them accept arguments. We used this to create a function to determine whether a product is in stock or not and disabled the add to cart button when it is not in stock, also refactoring the add item to cart function to use that getter.