Class: Vuex for everyone Chapter: 4 Lesson: 13 Title: Vuex map helpers Link: https://vueschool.io/lessons/vuex-mapping-state-mutations-getters-and-actions
In this lesson we learned how to use vuex map helpers for mapState, mapGetters and mapActions, replacing our long monotonous implementation with a more sleek one for accessing several properties that we were referencing through long chained property or function calls. This cleaned up the code quite a bit. Note, you can use an array so long as you do not need to rename a mapped property, but if you need to map a property to a different name you need to use an object. {newName: "existingName"}