Class: Vuex for everyone Chapter: 4 Lesson: 15 Title: Vuex modules Link: https://vueschool.io/lessons/vuex-modules
In this lesson we converted our vuex store over to multiple vuex modules, one for cart and one for products. The big takeaway from me was that you have to access rootState in some situations for modules, separately from local state. rootState is also called the global state. Overall this was pretty smooth, but it may take me some more tinkering to get a good grasp on when I would need to access rootState vs local in some cases but I think this would come to me pretty quickly with more use. We also looked in vue dev tools to see how the modules looked in state, and highlighted that getters/mutations/actions are still global.