Class: Vuex for everyone Chapter: 4 Lesson: 16 Title: Namespaced Vuex modules Link: https://vueschool.io/lessons/module-namespaces
In this lesson we converted our cart and products vuex modules over to namespaced modules. This makes managing multiple modules in a project easier as it prevents collisions of named functions in the modules. It was highlighted that if two modules each have a getter with the same function name as an example that they will both fire unless they are namespaced. It also helps organize the code a bit better, making it clearer what objects a particulare store should be interacting with. One of the interesting highlights was the need to use root: true within stores to properly use a named module within a different named module.